Home

CircularList.jl

CircularList.Buffer

CircularList.Node

Base.delete!

Base.insert!

Base.length

CircularList.allocate!

CircularList.backward!

CircularList.circularlist

CircularList.current

CircularList.forward!

CircularList.next

CircularList.previous

CircularList.shift!

Base.delete! — Method.

Delete current node and return the previous node.

Warning: removed nodes are not reclaimed from memory for simplicity reasons

source

Base.insert! — Method.

Insert a new node after the current node and return the new node.

source

Base.length — Method.

Returns the length of the circular list

source

CircularList.backward! — Method.

Shift the current pointer backward.

source

CircularList.circularlist — Method.

Create a circular list with the specified data element.

source

CircularList.current — Method.

Current node

source

CircularList.forward! — Method.

Shift the current pointer forward.

source

CircularList.next — Method.

Next node

source

CircularList.previous — Method.

Previous node

source

CircularList.shift! — Method.

Shift the current pointer forward or backward.

source

CircularList.Buffer — Type.

Buffer is used to hold a pre-allocated vector of nodes.

source

CircularList.Node — Type.

Double linked list

source

CircularList.allocate! — Method.

Allocates a new uninitialized node in the circular list

source