Home

CircularList.jl

CircularList.List

CircularList.Node

Base.delete!

Base.eltype

Base.insert!

Base.iterate

Base.length

Base.show

Base.show

Base.size

CircularList.allocate!

CircularList.backward!

CircularList.circularlist

CircularList.circularlist

CircularList.current

CircularList.forward!

CircularList.head

CircularList.next

CircularList.previous

CircularList.shift!

CircularList.tail

Base.delete! — Method.

Delete current node and return the previous node.

Warning: removed nodes are not reclaimed from memory for simplicity reasons

source

Base.eltype — Method.

Return the element type of the list

source

Base.insert! — Method.

Insert a new node after the current node and return the new node.

source

Base.iterate — Function.

Iteration protocol implementation.

source

Base.length — Method.

Returns the length of the circular list

source

Base.show — Method.

Show list.

source

Base.show — Method.

Show node

source

Base.size — Method.

Return the size of the list.

source

CircularList.backward! — Method.

Shift the current pointer backward.

source

CircularList.circularlist — Method.

Create a circular list from any vector

source

CircularList.circularlist — Method.

Create a circular list with the specified data element.

source

CircularList.current — Method.

Return the current node.

source

CircularList.forward! — Method.

Shift the current pointer forward.

source

CircularList.head — Method.

Return the head of the list (current node).

source

CircularList.next — Method.

Return the ext node.

source

CircularList.previous — Method.

Return the previous node.

source

CircularList.shift! — Method.

Shift the current pointer forward or backward.

source

CircularList.tail — Method.

Return the tail of the list (last node)

source

CircularList.List — Type.

List is used to hold a pre-allocated vector of nodes.

source

CircularList.Node — Type.

Doubly linked list implementation

source

CircularList.allocate! — Method.

Allocates a new uninitialized node in the circular list

source